work on bones, still mismatch in vertex buffer layout

+++ JonsEngine/src/Renderer/DirectX11/DX11Mesh.cpp
@@ -11,7 +11,7 @@ namespace JonsEngine
     const static uint32_t gVertexSize = sizeof(float) * 3;
     const static uint32_t gTangentSize = gVertexSize * 2;
     const static uint32_t gTexcoordSize = sizeof(float) * 2;
-    const static uint32_t gBoneIndexSize = sizeof(uint32_t) * MAX_BONES;
+    const static uint32_t gBoneIndexSize = sizeof(uint16_t) * MAX_BONES;
     const static uint32_t gBoneWeightSize = sizeof(float) * MAX_BONES;
     const static uint32_t gStaticOffset = 0;
     const std::array<uint16_t, 16> gAABBIndices = { 0, 1, 2, 3, 0, 5, 4, 3, 2, 7, 4, 7, 6, 5, 6, 1 };
